Module xcm_fuzzer::parachain::mock_msg_queue

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Re-exports

pub use __tt_error_token_1 as tt_error_token;

pub use __tt_default_parts_7 as tt_default_parts;

Structs

Pallet

The pallet implementing the on-chain logic.

Enums

Call

Contains one variant per dispatchable that can be called by an extrinsic.

Event

The event emitted by this pallet.

Traits

Config

Configuration trait of this pallet.

Store 🔒

Type Definitions

MessageId

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

ParachainId 🔒

ReceivedDmp 🔒

A queue of received DMP messages