Type Alias xcm_fuzzer::parachain::UncheckedExtrinsic

pub type UncheckedExtrinsic = UncheckedExtrinsic<MultiAddress<<<MultiSignature as Verify>::Signer as IdentifyAccount>::AccountId, ()>, RuntimeCall, MultiSignature, (CheckNonZeroSender<Runtime>,)>;

Aliased Type

struct UncheckedExtrinsic {
    pub signature: Option<(MultiAddress<AccountId32, ()>, MultiSignature, (CheckNonZeroSender<Runtime>,))>,
    pub function: RuntimeCall,
}

Fields
signature: Option<(MultiAddress<AccountId32, ()>, MultiSignature, (CheckNonZeroSender<Runtime>,))>

The signature, address, number of extrinsics have come before from the same signer and an era describing the longevity of this transaction, if this is a signed extrinsic.

None if it is unsigned or an inherent.

function: RuntimeCall

The function that should be called.