Module xcm_fuzzer::relay_chain

Relay chain runtime mock.

Structs

AnyNetwork

BaseXcmWeight

BlockHashCount

ExistentialDeposit

FirstMessageFactorPercent

KsmPerSecondPerByte

MaxAssetsIntoHolding

MaxInstructions

MaxLocks

MaxReserves

MessageProcessor

Message processor to handle any messages that were enqueued into the MessageQueue pallet.

MessageQueueHeapSize

MessageQueueMaxStale

MessageQueueServiceWeight

Amount of weight that can be spent per block to service messages.

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

Runtime

RuntimeGenesisConfig

RuntimeOrigin

The runtime origin type representing the origin of a call.

ThisNetwork

TokenLocation

UniversalLocation

XcmConfig

Enums

OriginCaller

RuntimeCall

RuntimeError

RuntimeEvent

RuntimeFreezeReason

A reason for placing a freeze on funds.

RuntimeHoldReason

A reason for placing a hold on funds.

RuntimeLockId

An identifier for each lock placed on funds.

RuntimeSlashReason

A reason for slashing funds.

Constants

LOG_TARGET 🔒

Traits

InherentDataExt 🔒

Functions

handle_put_error_result 🔒

Type Definitions

AccountId

AllPalletsDeprecated

All pallets included in the runtime as a nested tuple of types.

AllPalletsReversedWithSystemFirstDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order. With the system pallet first.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithSystemReversedDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AllPalletsWithoutSystemReversedDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order. Excludes the System pallet.

Balance

Balances

BalancesConfig

Barrier

Block 🔒

LocalAssetTransactor

LocalOriginConverter 🔒

LocalOriginToLocation

MessageQueue

ParasOrigin

SovereignAccountOf

System

SystemConfig

XcmPallet

XcmRouter

__hidden_use_of_unchecked_extrinsic 🔒