Module xcm_fuzzer::parachain

Parachain runtime mock.

Modules

mock_msg_queue

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Structs

BlockHashCount

ExistentialDeposit

KsmLocation

KsmPerSecondPerByte

MaxAssetsIntoHolding

MaxInstructions

MaxLocks

MaxReserves

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

RelayNetwork

ReservedDmpWeight

ReservedXcmpWeight

Runtime

RuntimeOrigin

The runtime origin type representing the origin of a call.

UnitWeightCost

UniversalLocation

XcmConfig

Enums

OriginCaller

RuntimeCall

RuntimeEvent

RuntimeFreezeReason

RuntimeHoldReason

RuntimeLockId

RuntimeSlashReason

Traits

InherentDataExt 🔒

Type Definitions

AccountId

AllPalletsDeprecated

All pallets included in the runtime as a nested tuple of types.

AllPalletsReversedWithSystemFirstDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order. With the system pallet first.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithSystemReversedDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AllPalletsWithoutSystemReversedDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order. Excludes the System pallet.

Balance

Balances

Barrier

Block 🔒

LocalAssetTransactor

LocalOriginToLocation

LocationToAccountId

MsgQueue

PolkadotXcm

System

UncheckedExtrinsic 🔒

XcmOriginToCallOrigin

XcmRouter