xcm_fuzzer

Module parachain

Parachain runtime mock.

Modules

mock_msg_queue

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

Structs

BlockHashCount

ExistentialDeposit

KsmLocation

KsmPerSecondPerByte

MaxAssetsIntoHolding

MaxInstructions

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

RelayNetwork

ReservedDmpWeight

ReservedXcmpWeight

Runtime

RuntimeGenesisConfig

RuntimeOrigin

The runtime origin type representing the origin of a call.

UnitWeightCost

UniversalLocation

XcmConfig

Enums

OriginCaller

RuntimeCall

The aggregated runtime call type.

RuntimeError

RuntimeEvent

RuntimeFreezeReason

A reason for placing a freeze on funds.

RuntimeHoldReason

A reason for placing a hold on funds.

RuntimeLockId

An identifier for each lock placed on funds.

RuntimeSlashReason

A reason for slashing funds.

RuntimeTask

An aggregation of all Task enums across all pallets included in the current runtime.

RuntimeViewFunction

Runtime query type.

Traits

InherentDataExt ๐

Type Aliases

AccountId

Address

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

Balance

Balances

BalancesConfig

Barrier

Block

BlockNumber

Header

LocalAssetTransactor

LocalOriginToLocation

LocationToAccountId

MsgQueue

PolkadotXcm

PolkadotXcmConfig

Signature

System

SystemConfig

TxExtension

UncheckedExtrinsic

XcmOriginToCallOrigin

XcmRouter