Module xcm_fuzzer::parachain

Parachain runtime mock.

Modules

mock_msg_queue

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Structs

Ancestry

BlockHashCount

ExistentialDeposit

KsmLocation

KsmPerSecond

MaxInstructions

MaxLocks

MaxReserves

Origin

The runtime origin type represanting the origin of a call.

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

RelayNetwork

ReservedDmpWeight

ReservedXcmpWeight

Runtime

UnitWeightCost

XcmConfig

Enums

Call

Event

OriginCaller

Traits

InherentDataExt

Type Definitions

AccountId

AllModulesDeprecated

All modules included in the runtime as a nested tuple of types. Excludes the System pallet.

AllModulesWithSystemDeprecated

All modules included in the runtime as a nested tuple of types.

AllPallets

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

Balance

Balances

Barrier

Block

LocalAssetTransactor

LocalOriginToLocation

LocationToAccountId

MsgQueue

PolkadotXcm

System

UncheckedExtrinsic

XcmOriginToCallOrigin

XcmRouter