Module xcm_fuzzer::relay_chain

Relay chain runtime mock.

Structs

AnyNetwork

BaseXcmWeight

BlockHashCount

ExistentialDeposit

FirstMessageFactorPercent

KsmPerSecondPerByte

MaxAssetsIntoHolding

MaxInstructions

MaxLocks

MaxReserves

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

Runtime

RuntimeOrigin

The runtime origin type representing the origin of a call.

ThisNetwork

TokenLocation

UniversalLocation

XcmConfig

Enums

OriginCaller

RuntimeCall

RuntimeEvent

RuntimeFreezeReason

RuntimeHoldReason

RuntimeLockId

RuntimeSlashReason

Traits

InherentDataExt 🔒

Type Definitions

AccountId

AllPalletsDeprecated

All pallets included in the runtime as a nested tuple of types.

AllPalletsReversedWithSystemFirstDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order. With the system pallet first.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithSystemReversedDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AllPalletsWithoutSystemReversedDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order. Excludes the System pallet.

Balance

Balances

Barrier

Block 🔒

LocalAssetTransactor

LocalOriginConverter 🔒

LocalOriginToLocation

ParasOrigin

ParasUmp

SovereignAccountOf

System

UncheckedExtrinsic 🔒

XcmPallet

XcmRouter