xcm_fuzzer

Module relay_chain

Relay chain runtime mock.

Structs

AnyNetwork

BaseXcmWeight

BlockHashCount

ExistentialDeposit

KsmPerSecondPerByte

MaxAssetsIntoHolding

MaxInstructions

MessageProcessor

Message processor to handle any messages that were enqueued into the MessageQueue pallet.

MessageQueueHeapSize

MessageQueueMaxStale

MessageQueueServiceWeight

Amount of weight that can be spent per block to service messages.

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

Runtime

RuntimeGenesisConfig

RuntimeOrigin

The runtime origin type representing the origin of a call.

ThisNetwork

TokenLocation

UniversalLocation

XcmConfig

Enums

OriginCaller

RuntimeCall

The aggregated runtime call type.

RuntimeError

RuntimeEvent

RuntimeFreezeReason

A reason for placing a freeze on funds.

RuntimeHoldReason

A reason for placing a hold on funds.

RuntimeLockId

An identifier for each lock placed on funds.

RuntimeSlashReason

A reason for slashing funds.

RuntimeTask

An aggregation of all Task enums across all pallets included in the current runtime.

RuntimeViewFunction

Runtime query type.

Traits

InherentDataExt 🔒

Type Aliases

AccountId

Address

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

Balance

Balances

BalancesConfig

Barrier

Block

BlockNumber

Header

LocalAssetTransactor

LocalOriginConverter 🔒

LocalOriginToLocation

MessageQueue

ParasOrigin

Signature

SovereignAccountOf

System

SystemConfig

TxExtension

UncheckedExtrinsic

XcmPallet

XcmPalletConfig

XcmRouter