Module xcm_fuzzer::relay_chain

Relay chain runtime mock.

Structs

Ancestry

AnyNetwork

BaseXcmWeight

BlockHashCount

ExistentialDeposit

FirstMessageFactorPercent

KsmLocation

KsmPerSecond

KusamaNetwork

MaxInstructions

MaxLocks

MaxReserves

Origin

The runtime origin type represanting the origin of a call.

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

Runtime

UnitWeightCost

XcmConfig

Enums

Call

Event

OriginCaller

Traits

InherentDataExt

Type Definitions

AccountId

AllModulesDeprecated

All modules included in the runtime as a nested tuple of types. Excludes the System pallet.

AllModulesWithSystemDeprecated

All modules included in the runtime as a nested tuple of types.

AllPallets

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

Balance

Balances

Barrier

Block

LocalAssetTransactor

LocalOriginConverter

LocalOriginToLocation

ParasOrigin

ParasUmp

SovereignAccountOf

System

UncheckedExtrinsic

XcmPallet

XcmRouter